:sparkles: Feat: 장바구니 아이템 수정 API 연동

+++ app/src/main/java/com/ite/sws/domain/cart/data/CartItem.kt
@@ -13,7 +13,7 @@ package com.ite.sws.domain.cart.data
  * </pre>
  */
 data class CartItem(
-    val productId: Int,
+    val productId: Long,
     val quantity: Int,
     val productName: String,
     val productPrice: Int,
